refactor: Move rearrange and move peers folder into individual folders
