--- crates/libtortillas/src/peers/peer/mod.rs
@@ -1,4 +1,0 @@
-pub mod id;
-pub mod info;
-pub mod state;
-pub mod supports;
